% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{apply_random_sine_wave}
\alias{apply_random_sine_wave}
\title{Apply a random sine wave}
\usage{
apply_random_sine_wave(x, num_sine_waves, frequency, seed, print_seed = FALSE)
}
\arguments{
\item{x}{a vector of x values}

\item{num_sine_waves}{number individual sine waves to sum}

\item{frequency}{a numeric value, frequency will be selected randomly between
the - and + of the value specified in this argument}

\item{seed}{a numeric value, the seed to initialize psuedo-randomness}

\item{print_seed}{a boolean value, set to TRUE to print seed}
}
\value{
a vector of y values for the random sine wave
}
\description{
Apply a random sine wave
}
