% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric.R
\name{create_voronoi}
\alias{create_voronoi}
\title{Generate random points & associated varanoi}
\usage{
create_voronoi(n_points, distribution)
}
\arguments{
\item{n_points}{an integer value, number of points}

\item{distribution}{'beta', 'uniform', or 'normal'}
}
\value{
a random work of art as a ggplot object
}
\description{
Generate random points & associated varanoi
}
