% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric.R
\name{generate_fabric}
\alias{generate_fabric}
\title{Generate fabric of reality}
\usage{
generate_fabric(n_points, distribution = "beta", seed)
}
\arguments{
\item{n_points}{an integer value, number of points}

\item{distribution}{'beta', 'uniform', or 'normal'}

\item{seed}{a numeric value, the seed to initialize psuedo-randomness}
}
\value{
a random work of art as a ggplot object
}
\description{
Generate fabric of reality
}
