% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{generate_inner_circles}
\alias{generate_inner_circles}
\title{Generate inner concentric circles}
\usage{
generate_inner_circles(
  x0,
  y0,
  r,
  color,
  num_inner_circles = 10,
  equal_spacing = TRUE
)
}
\arguments{
\item{x0}{a numeric value, the x-coordinate of the center of the source circle}

\item{y0}{a numeric value, the y-coordinate of the center of the source circle}

\item{r}{a numeric value, the radius of the source circle}

\item{color}{a character value, the color (hex code) of the source circle}

\item{num_inner_circles}{an integer value, the desired number of inner concentric
circles}

\item{equal_spacing}{a boolean value, if true the spacing betwen concentric
circles will be equal}
}
\value{
a tibble contain with one row per concentric circle, including the
original source circle, including x0, y0, r, and color
}
\description{
Generate inner concentric circles
}
