% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_and_polygons.R
\name{generate_lines_and_polygons}
\alias{generate_lines_and_polygons}
\title{Generate lines and polygons}
\usage{
generate_lines_and_polygons(num_shapes, background = "black", seed)
}
\arguments{
\item{num_shapes}{an integer value, number of points}

\item{background}{'white' or 'black'}
}
\value{
a random work of art as a ggplot object
}
\description{
Generate lines and polygons
}
