% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_and_rectangles.R
\name{generate_lines_and_rectangles}
\alias{generate_lines_and_rectangles}
\title{Generate lines and rectangles}
\usage{
generate_lines_and_rectangles(
  num_lines,
  p_v_line,
  max_line_size,
  num_rect,
  max_rect_size,
  seed
)
}
\arguments{
\item{num_lines}{an integer value >=0, the number of lines in the artwork;
if unspecified, a value will be chosen randomly}

\item{p_v_line}{a numeric value between 0 and 1, the probability that each
line will be vertical as opposed to horizontal; if unspecified, a value will
be chosen randomly}

\item{max_line_size}{a numeric value between 0 and 1, the maximum size of
each line; if unspecified, a value will be chosen randomly}

\item{num_rect}{an integer value >=0, the number of rectangles in the artwork;
if unspecified, a value will be chosen randomly}

\item{max_rect_size}{a numberic value between 0 and 1, the maximum size of
each rectangle; if unspecified, a value will be chosen randomly}

\item{seed}{a numeric value, the seed to initialize psuedo-randomness}
}
\value{
a random work of art as a ggplot object
}
\description{
Generate lines and rectangles
}
