% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{generate_panels}
\alias{generate_panels}
\title{Generate panels}
\usage{
generate_panels(num_panels)
}
\arguments{
\item{num_panels}{the number of panels in the artwork; if unspecified, a value
will be chosen randomly}
}
\value{
a random work of art as a ggplot object
}
\description{
Generate panels
}
