% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perception_matrix.R
\name{generate_perception_matrix}
\alias{generate_perception_matrix}
\title{Generate perception matrix}
\usage{
generate_perception_matrix(
  source_num_sine_waves,
  source_frequency,
  num_waves,
  even_spacing = TRUE,
  seed
)
}
\arguments{
\item{source_num_sine_waves}{an integer value, the number of sine waves
that will be summed to generate the waves of the matrix (this is used as
input to the apply_random_sine_wave() function)}

\item{source_frequency}{a numeric value, controls maximum frequency of
waves of the matrix (this is used as input to apply_random_sine_wave()
function)}

\item{num_waves}{an integer value, the approximate number of waves that make
up the matrix in each directional axis}

\item{even_spacing}{a boolean value, set to TRUE to make the waves evenly
spaced}

\item{seed}{a numeric value, the seed to initialize psuedo-randomness}
}
\value{
a random work of art as a ggplot object
}
\description{
Generate perception matrix
}
