% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_and_polygons.R
\name{generate_points_and_edges}
\alias{generate_points_and_edges}
\title{Generate lines and polygons}
\usage{
generate_points_and_edges(n_points, n_edges)
}
\arguments{
\item{n_points}{an integer value, number of points}

\item{n_edges}{an integer value, number of edges for each point}
}
\value{
a tibble of edges to support the generate_lines_and_polygons()
function
}
\description{
Generate lines and polygons
}
