% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_waves.R
\name{generate_string_waves}
\alias{generate_string_waves}
\title{Generate string waves}
\usage{
generate_string_waves(
  num_waves,
  num_colors,
  wobbliness,
  frequency,
  stability_n,
  stability_sd,
  loess_span,
  anchor = F,
  frame = "square",
  seed
)
}
\arguments{
\item{num_waves}{an integer value >=1, the number of waves desired}

\item{wobbliness}{an integer value >=1, higher values will make the waves
more wobbly (i.e. having peaks a trough of variable height); technically
this is the number if individual sine functions that are summed together to get
the final wave function}

\item{frequency}{a numberic value >= 0, higher values will increase the frequency
of the wave function; technically this is the passed to the frequency argument
of the apply_random_sine_wave() function (see apply_random_sine_wave() for details)}

\item{stability_n}{an integer value >=1, higher values will make the waves more
stable; technically, this is the number of data points sampled to generate
each wave}

\item{stability_sd}{a numeric value >=0, lower values will make the waves
more stable; technically, this is the standard deviation of the error applied
to the output of each wave function}

\item{loess_span}{a numeric value between 0-1, lower numbers will make each wave
more "wiggly" when fitting the underlying points, values < 0.3 seem to throw an
error}

\item{anchor}{a boolean value, set to true to make the waves converge to
anchor points}

\item{frame}{'points', 'circle', 'square', or 'none'; specifies the frame desired for
each artwork}

\item{seed}{a numeric value, the seed to initialize psuedo-randomness}
}
\value{
a random work of art as a ggplot object
}
\description{
Generate string waves
}
