% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{generate_web}
\alias{generate_web}
\title{Generate webs}
\usage{
generate_web(n = 50, seed)
}
\arguments{
\item{n}{an integer value, number of points}

\item{seed}{a numeric value, the seed to initialize psuedo-randomness}
}
\value{
a random work of art as a ggplot object
}
\description{
Generate webs
}
