% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_random_color_pool}
\alias{get_random_color_pool}
\title{Get a random color pool}
\usage{
get_random_color_pool(num_colors)
}
\arguments{
\item{num_colors}{number of colors desired}
}
\value{
a vector of adjacent hex codes
}
\description{
Get a random color pool
}
