% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{rotate_points}
\alias{rotate_points}
\title{Rotate points}
\usage{
rotate_points(x_y_df, angle, seed, print_seed = FALSE)
}
\arguments{
\item{x_y_df}{a dataframe/tibble with columns x & y}

\item{angle}{a numeric value, specified angle of rotation in radians}

\item{seed}{a numeric value, the seed to initialize psuedo-randomness}

\item{print_seed}{a boolean value, set to TRUE to print seed}
}
\value{
a vector of y values for the random sine wave
}
\description{
Rotate points
}
