% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{generate_panels}
\alias{generate_panels}
\title{Generate panels}
\usage{
generate_panels(num_panels, square = TRUE, single_color = TRUE, seed)
}
\arguments{
\item{num_panels}{the number of panels in the artwork; if unspecified, a value
will be chosen randomly}

\item{square}{a boolean value, set to TRUE for the artwork to be square}

\item{single_color}{a boolean value, set to TRUE to use a single color}

\item{seed}{a numeric value, the seed to initialize psuedo-randomness}
}
\value{
a random work of art as a ggplot object
}
\description{
Generate panels
}
