% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_waves.R
\name{generate_string_waves}
\alias{generate_string_waves}
\title{Generate string waves}
\usage{
generate_string_waves(
  num_waves,
  stability_n,
  stability_sd,
  wobbliness,
  frequency,
  frame = "points"
)
}
\arguments{
\item{stability_n}{an integer value >=1, higher values will make the waves more
stable; technically, this is the number of data points sampled to generate
each wave}

\item{stability_sd}{a numeric value >=0, lower values will make the waves
more stable; technically, this is the standard deviation of the error applied
to the output of each wave function}

\item{wobbliness}{an integer value >=1, higher values will make the waves
more wobbly (i.e. having peaks a trough of variable height); technically
this is the number if individual sine functions that are summed together to get
the final wave function}

\item{frequency}{a numberic value >= 0, higher values will increase the frequency
of the wave function; technically this is the passed to the frequency argument
of the random_sine_wave() function (see random_sine_wave() for details)}

\item{frame}{'points', 'circle' or 'none'; specifies the frame desired for
each artwork; defaults to 'points' because they look kind of cool.}
}
\value{
a random work of art as a ggplot object
}
\description{
Generate string waves
}
